% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-getvarnames.R
\name{getvarnames}
\alias{getvarnames}
\title{Internal helper: get variable names from a model call}
\usage{
getvarnames(call)
}
\description{
Internal helper: get variable names from a model call
}
\keyword{internal}
